+++ icegraph/exceptions/__init__.py
@@ -1,8 +1,9 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from .models import IceGraphError
+from .models import IceGraphError, IceCubeImportError
 
 IceGraphError.__module__ = __name__
+IceCubeImportError.__module__ = __name__
 
-__all__ = ["IceGraphError"]
+__all__ = ["IceGraphError", "IceCubeImportError"]