+++ icegraph/renderer/__init__.py
@@ -1,8 +1,11 @@
 # Copyright (c) 2025 University of Maryland and the IceCube Collaboration.
 # Developed by Taylor St Jean
 
-from .models import FeaturePlot
+from .models import FeaturePlot, CDFPlot, PDFPlot, ChargeDistributionPlot
 
 FeaturePlot.__module__ = __name__
+CDFPlot.__module__ = __name__
+PDFPlot.__module__ = __name__
+ChargeDistributionPlot.__module__ = __name__
 
-__all__ = ["FeaturePlot"]
+__all__ = ["FeaturePlot", "CDFPlot", "PDFPlot", "ChargeDistributionPlot"]