+++ icegraph/renderer/models.py
@@ -2,22 +2,89 @@
 # Developed by Taylor St Jean
 
 from pathlib import Path
+from typing import Optional
 
 import numpy as np
-import matplotlib.pyplot as plt
+import plotly.graph_objects as go
 
-from .base import IGPlot
+from .base import IGPlot, IGDistributionPlot
 from icegraph.data.processor import generate_vector_mapping
+from icegraph.data.pulses import Pulses
+from icegraph.exceptions import IceCubeImportError
 
-__all__ = ["FeaturePlot"]
+try:
+    from icecube.icetray import OMKey as _OMKey
+except ImportError:
+    _OMKey = IceCubeImportError.IceCubeMissingBase
+
+OMKey = _OMKey
+
+__all__ = ["FeaturePlot", "CDFPlot", "PDFPlot", "ChargeDistributionPlot"]
 
 
-class CDFPlot(IGPlot):
+class CDFPlot(IGDistributionPlot):
 
-    subplots = (1, 1)
+    plot_type = "CDF"
 
-    def plot(self, feature: str, save_path: Path | None=None) -> None:
-        pass
+    def _populate_plot(self, target_dom: Optional[OMKey] = None) -> Pulses.DOMMetadata:
+        times, cdf, metadata = self._pulses.get_cdf(target_dom)
+
+        self._fig.add_trace(go.Scatter(x=times, y=cdf, mode="lines", name="CDF", line=dict(width=4)))
+        self._fig.update_layout(
+            xaxis_title="Time [ns]",
+            yaxis_title="Normalized Cumulative Charge"
+        )
+
+        return metadata
+
+
+class PDFPlot(IGDistributionPlot):
+
+    plot_type = "PDF"
+
+    def _populate_plot(self, target_dom: Optional[OMKey] = None) -> Pulses.DOMMetadata:
+        times, charges, metadata = self._pulses.get_pulses(target_dom)
+
+        # Compute weighted histogram
+        min_t = np.floor(min(times))
+        max_t = min_t + 500  # avoid any noise and cut to 0.5 microsecond
+        bin_edges = np.linspace(min_t, max_t, 51)
+        hist_vals, bins = np.histogram(times, bins=bin_edges, weights=charges)
+
+        bin_centers = 0.5 * (bins[1:] + bins[:-1])
+
+        self._fig.add_trace(go.Bar(
+            x=bin_centers,
+            y=hist_vals,
+            width=np.diff(bins),
+            name='PDF'
+        ))
+
+        self._fig.update_layout(
+            xaxis_title="Time [ns]",
+            yaxis_title="Charge [PE]",
+            bargap=0.1
+        )
+
+        return metadata
+
+
+class ChargeDistributionPlot(IGDistributionPlot):
+
+    plot_type = "charge_dist"
+
+    def _populate_plot(self, target_dom: Optional[OMKey] = None) -> Pulses.DOMMetadata:
+        _, charges, metadata = self._pulses.get_pulses(target_dom)
+
+        self._fig.add_trace(go.Histogram(x=charges, nbinsx=100, name="CD"))
+        self._fig.update_layout(
+            xaxis_title="Charge [PE]",
+            yaxis_title="Count",
+            bargap=0.1
+        )
+
+        return metadata
+
 
 class FeaturePlot(IGPlot):
     """
@@ -35,7 +102,7 @@ class FeaturePlot(IGPlot):
             save_path (Path): Path to save the plot. Defaults to output_dir specified in the global config file.
         """
         if not save_path:
-            save_path = Path(self._config.user_config.io.output_dir) / f"feature_plot_{feature}.png"
+            save_path = Path(self._config.user_config.io.default_dir) / f"feature_plot_{feature}.png"
 
         # label axes
         self._ax[0].set_xlabel("dom_x")